rename all files and directory to conform to PEP 8
