+++ perform/higher_order_funcs.py
