+++ perform/rom/projection_rom/autoencoder_proj_rom/autoencoder_tfkeras/autoencoder_galerkin_proj_tfkeras.py
@@ -2,7 +2,7 @@ import numpy as np
 import tensorflow as tf
 from scipy.linalg import pinv
 
-from perform.rom.projectionROM.autoencoderProjROM.autoencoderTFKeras.autoencoderTFKeras import AutoencoderTFKeras
+from perform.rom.projection_rom.autoencoder_proj_rom.autoencoder_tfkeras.autoencoder_tfkeras import AutoencoderTFKeras
 
 
 class AutoencoderGalerkinProjTFKeras(AutoencoderTFKeras):
