+++ perform/rom/projection_rom/autoencoder_proj_rom/autoencoder_tfkeras/autoencoder_lspg_proj_tfkeras.py
@@ -1,7 +1,7 @@
 import numpy as np
 import tensorflow as tf
 
-from perform.rom.projectionROM.autoencoderProjROM.autoencoderTFKeras.autoencoderTFKeras import AutoencoderTFKeras
+from perform.rom.projection_rom.autoencoder_proj_rom.autoencoder_tfkeras.autoencoder_tfkeras import AutoencoderTFKeras
 
 
 class AutoencoderLSPGProjTFKeras(AutoencoderTFKeras):
