+++ perform/rom/projection_rom/linear_proj_rom/linear_splsvt_proj.py
@@ -1,39 +1,38 @@
 import numpy as np
 
-from perform.rom.projectionROM.linearProjROM.linearProjROM import LinearProjROM
+from perform.rom.projection_rom.linear_proj_rom.linear_proj_rom import LinearProjROM
 
 
-class LinearLSPGProj(LinearProjROM):
+class LinearSPLSVTProj(LinearProjROM):
 	"""
-	Class for linear decoder and least-squares Petrov-Galerkin projection
+	Class for linear decoder and SP-LSVT formulation
 	Trial basis is assumed to represent the conserved variables
 	"""
 
 	def __init__(self, model_idx, rom_domain, solver, sol_domain):
 
-		# I'm not going to code LSPG with explicit time integrator,
-		# 	it's a pointless exercise
 		if rom_domain.time_integrator.time_type == "explicit":
-			raise ValueError("LSPG with an explicit time integrator deteriorates"
-							+ " to Galerkin, please use Galerkin or select an"
-							+ "implicit time integrator.")
+			raise ValueError("Explicit SP-LSVT not implemented yet")
 
-		if rom_domain.time_integrator.dual_time:
-			raise ValueError("LSPG is intended for conservative variable evolution,"
-							+ "please set dual_time = False")
+		if ((rom_domain.time_integrator.time_type == "implicit")
+				and (not rom_domain.time_integrator.dual_time)):
+			raise ValueError("SP-LSVT is intended for primitive variable"
+							+ "evolution, please use Galerkin or LSPG,"
+							+ " or set dual_time = True")
 
 		super().__init__(model_idx, rom_domain, solver, sol_domain)
 
 	def calc_d_code(self, res_jacob, res, sol_domain):
 		"""
-		Compute change in low-dimensional state for implicit scheme Newton iteration
+		Compute change in low-dimensional state for implicit scheme
+		Newton iteration
 		"""
 
 		# TODO: add hyper-reduction
 
 		# TODO: scaled_trial_basis should be calculated once
 		scaled_trial_basis = \
-			self.trial_basis * self.norm_fac_prof_cons.ravel(order="C")[:, None]
+			self.trial_basis * self.norm_fac_prof_prim.ravel(order="C")[:, None]
 
 		# compute test basis
 		test_basis = (
